chore: moving some files and making stuff pretty
