+++ include/core/Reactor.hpp
@@ -4,7 +4,7 @@
 
 #include "core/Dispatcher.hpp"
 #include "core/IHandler.hpp"
-#include "http/VirtualServer.hpp"
+#include "core/VirtualServer.hpp"
 #include "http/constants.hpp"
 #include "shared/NonCopyable.hpp"
