+++ src/core/IOHandler.cpp
@@ -16,6 +16,7 @@ namespace core {
 	}
 
 	void IOHandler::handle(int32_t, uint32_t) {
+		std::cout << "handling" << std::endl;
 	}
 
 } /* namespace core */
